feat: Extracted handleSubmit to page.tsx
